Falta proyectos

+++ views/modules/indicator_mod/indicator_manage/main.indicator.php
@@ -24,13 +24,10 @@
                                             <td><?php echo $row["ind_categoria"]; ?></td>
                                             <td><?php echo $row["ind_responsable"]; ?></td>
                                             <td>
-                                                <a href="" class="icon-table">
-                                                    <i class="material-icons icon-green">remove_red_eye</i>
-                                                </a>
                                                 <a href="index.php?c=indicator&a=update&token=<?php echo $row['ind_code'];?>" class="icon-table">
                                                     <i class="material-icons icon-yellow">create</i>
                                                 </a>
-                                                <a href="index.php?c=indicator&a=delete&token=<?php echo $row['ind_code'];?>" class="icon-table">
+                                                <a href="index.php?c=indicator&a=delete&token=<?php echo $row['ind_code'];?>" class="icon-table" onclick="return confirm('Está seguro?');">
                                                     <i class="material-icons icon-red">delete</i>
                                                 </a>
                                             </td>
